fix too many rernders (#103672)
